Add Travis CI builds for more recent DMD releases
This patch adds builds for all the most recent patch releases of DMD, up
to 2.085.1.  More recent versions than this cannot be included for now
due to a breaking change in DMD 2.086.0 that causes a failure with more
recent releases of `taggedalgebraic`:
https://github.com/dlang-community/std_data_json/issues/44#issuecomment-571111135

Once that issue is fixed it should be possible to include all the most
recent compiler versions.

LDC versions have been left untouched for now: it seems best to upgrade
those once compatibility with the most recent DMD has been guaranteed,
and we can decide how many historical versions we want to support.