Add multiple Travis scripts to run with and without dub upgrade
Since `dub.selections.json` is fixed in the repo, CI will fail to pick
up on bugs associated with changes in more recent dependency versions,
such as the current known issue with `taggedalgebraic` >= 0.11.0 and
DMD >= 2.086.0:
https://github.com/dlang-community/std_data_json/issues/44#issuecomment-571111135

This patch adds custom `script` settings, with two entries:

  * a script which runs the default `dub test` call, which will use
    the dependencies specified in `dub.selections.json`

  * a script which calls `dub upgrade` before the `dub test` call,
    ensuring that the latest dependency versions are used

This should ensure that CI validates that `stdx.data.json` is compatible
with both its lowest and highest stated dependency versions.

Note that no new build matrix entries are added: both scripts are run in
the course of a single build.

This is expected to cause a failure for the `dmd-beta` build, since any
DMD >= 2.086.0 triggers the aforementioned `taggedalgebraic` issue, but
since these are `allowed_failures` in Travis config CI as a whole should
still pass (these failures are effectively treated as warnings of future
breaking changes rather than as errors).