Add more extensive unittests for JSONValue equality comparisons
`JSONValue` equality comparison relies on `TaggedAlgebraic.opEquals` as
accessed via `alias payload this`.  Unfortunately this breaks in several
ways for taggedalgebraic >= 0.11.0, in some cases for all DMD versions,
and in others only for DMD >= 2.086.0.

To fix these issues we must implement an explicit `JSONValue.opEquals`
method.  Before doing this, it seems a good idea to put in place some
unittests to ensure that we do not break any pre-existing behaviour.
This patch introduces a comprehensive set of tests to cover all the
default-enabled payload types: null, bool, integer, floating point,
string, array, and object.

Tests for `BigInt` are not included, as it is not obvious how best to
instantiate or compare this data: `JSONValue(BigInt("..."))` does not
work, while direct comparison of a `WrappedBigInt` will most likely
compare only pointers.  In any case, support for this payload type is
disabled by default, so identifying and fixing its issues is left for
future work.

Besides cases which fail with taggedalgebraic v0.11.x there are also a
number of array comparisons which fail regardless of the taggedalgebraic
version.  All such failure cases are included as commented-out `assert`
statements, ready to be uncommented when we add the explicit `opEquals`.

Part of https://github.com/dlang-community/std_data_json/issues/44.
Part of https://github.com/dlang-community/std_data_json/issues/48.