Add an explicit JSONValue.opEquals method to fix taggedalgebraic support
The original `JSONValue` implementation relied on `alias payload this`
for equality comparisons.  This breaks with taggedalgebraic >= 0.11.0,
with some of the failure cases emerging for all DMD versions, and others
only for DMD >= 2.086.0.

The former breakages seem to be related to changes in the implementation
of `TaggedAlgebraic.opEquals`, while the cases that emerge only for DMD
>= 2.086.0 appear to be related to a breaking change to prefer generated
struct `opEquals` over aliased `opEquals`:
https://dlang.org/changelog/2.086.0.html#alias_this_opEquals

The aliased `opEquals` also fails in a number of array comparison cases,
regardless of taggedalgebraic or compiler version.

This patch fixes all these issues by implementing an explicit `opEquals`
method.  As with the previous implementation this relies only on payload
and not on location, which appears to be in keeping with the rest of the
library.

We need to special case when `null` input is provided, as this could be
ambiguous as to which `PayloadUnion` entry to compare to.  Similarly, we
special-case comparison of `JSONValue` to `JSONValue` in order to ensure
that comparison is correctly based on their respective payloads.  In all
other cases, a direct comparison of input to payload appears sufficient.

Fixes https://github.com/dlang-community/std_data_json/issues/44 (*).
Fixes https://github.com/dlang-community/std_data_json/issues/48.

(* Proof that the DMD >= 2.086.0 issues are fixed should be visible via
the `dmd-beta` CI build, which should now pass.)